backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.2
pytest==7.4.0 